✨feat($front): 좋아요 구현

See merge request frozen_pig/content_i_like!113

+++ src/main/java/com/content_i_like/repository/LikesRepository.java
@@ -14,6 +14,6 @@ public interface LikesRepository extends JpaRepository<Likes, Long> {
   Optional<Likes> findLikesByMemberAndRecommend(Member member, Recommend post);
 
   @Query("select count(l) from Likes l where l.deletedAt is null and l.recommend = :recommend")
-  Integer countLikesByRecommend(@Param("recommend") Recommend recommend);
+  Long countLikesByRecommend(@Param("recommend") Recommend recommend);
 
 }
